move mosip utils to mosip.ts
